Main code
Briefly, we tried to fulfill the mediated schema with data from different sources XML sources based on the similarity.
The 3 XML sources are converted from the main 3 data sources CSV files ( about countries ).
The next step was making necessary customizations to the XML file to align it with our specific requirements and then starting parsing the XML files.
After that the task was subdivided is 3 main parties :
1-Matchers:  In order to find the similarity between mediated schema and data sources(XML.files) we used 2 different strategies (Full Path Strategy , Parent/Child Strategy)  in order to compare the results and increase the efficiency of matches.
We used 3 different Matchers:  - The First matcher Semantic similarity
                                                     -The second match Jaccard similarity
                                                     -The third matcher Levenshtein distance similarity

2-Combiners and Selectors: For each strategy, we iterate through the rows after combining the three matchers together into the same data frame with
the same column’s distribution and then pick up the average/max/min depending on the respective combiner l. then we pass to the selection which based on this 2 steps:
• Thresholding: we have created the function apply_threshold(df, threshold), that takes the result of any combiner and with a preset threshold as parameters and it iterates through the data frame. If the value >= to the threshold then we set the value to 1. Else it will be counted as 0.
• Cardinality: we have created the apply_cardinality_one_to_one(df): It takes the thresholded table as parameter, It iterates over each column and then find the max value in the respective index, then it stores the name if this column and its greatest element, so as the name of the row, then it change max value of the column to 100. 

3-Evaluation Metrics ( Precision , Recall , F1 score)  : Finally after getting the selected data source, we create a ground truth matrix in order to calculate the evaluation
metrics.